refactor(all): Completely restructure folder structure...files not yet updated
