+++ archived/hero-communicator-previous.js
