+++ archived/hero-communicator.js
