--- azureScripts/README.md
@@ -1,2 +1,0 @@
-javascript-battle-webworkers
-============================
