--- azureScripts/createDocker.README
@@ -1,35 +1,0 @@
-To create a docker image:
-
-sudo docker build -t nameofdockerimage folderwithdockerfile
-
-
-To spin up a container:
-
-sudo docker run -d -p 50013:8080 imageName
-
-50013 is any port number
-
-
-List containers
-
-sudo docker ps -a (all containers)
-sudo docker ps (running containers)
-sudo docker ps -l (last container)
-
-Stop a container
-
-sudo docker stop containerNameOrID
-
-List Images
-
-sudo docker images
-
-
-Remove Images
-
-sudo docker rmi imageName
-
-
-Terminal in container
-
-sudo docker attach containerNameOrId
