--- azureScripts/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "workers",
-  "version": "0.0.0",
-  "description": "javascript-battle-webworkers ============================",
-  "main": "Board.js",
-  "scripts": {
-    "test": "npm run test-node",
-    "test-node": "mocha -R spec ./test/spec/GameSpec.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/gjtrowbridge/javascript-battle-webworkers.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/gjtrowbridge/javascript-battle-webworkers/issues"
-  },
-  "homepage": "https://github.com/gjtrowbridge/javascript-battle-webworkers",
-  "dependencies": {
-    "body-parser": "^1.6.2",
-    "form-data": "^0.1.4",
-    "mongodb": "^1.4.8",
-    "multer": "^0.1.3",
-    "q": "^1.0.1",
-    "request": "^2.40.0",
-    "unirest": "^0.2.7"
-  }
-}
