+++ docker/container_interaction/communicate-with-container-functions.js
