+++ docker/container_interaction/start-hero-brain-container.sh
