+++ game_logic/game_classes/DiamondMine.js
