+++ game_logic/game_classes/Impassable.js
