+++ game_logic/maps/diamondMap.txt
