+++ secrets-template.js
